Got most of glslang working
Former-commit-id: 6bab83c5120c0ea64ebf5d722b9cce504f3cbe7c
Former-commit-id: bfbee0df82dc1eb47f8d02fbe9e970b277f00e33

+++ Sources/FlounderEngine/Meshes/Animations/MeshAnimated.cpp
@@ -104,7 +104,7 @@ namespace Flounder
 
 	void MeshAnimated::AddJointsToArray(const Joint &headJoint, std::vector<Matrix4*> *jointMatrices)
 	{
-		if (headJoint.GetIndex() < jointMatrices->size())
+		if (headJoint.GetIndex() < (int) jointMatrices->size())
 		{
 			jointMatrices->at(headJoint.GetIndex()) = headJoint.GetAnimatedTransform();
 		}
